refactor: 구조를 잘못잡아 다시 리펙토링
