+++ src/test/java/study/SetTest.java
@@ -1,4 +1,4 @@
-package study.step1;
+package study;
 
 import static org.assertj.core.api.Assertions.assertThat;
