+++ src/test/java/study/StringTest.java
@@ -1,4 +1,4 @@
-package study.step1;
+package study;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
